extract ConsoleRenderer to subdir
