+++ lib/console_tree_renderer/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ConsoleTreeRenderer
+  VERSION = "0.1.0"
+end
