--- lib/tree_renderer/version.rb
@@ -1,3 +1,0 @@
-module TreeRenderer
-  VERSION = "0.1.0"
-end
